requirements.txt
amqp==2.5.2
backcall==0.1.0
beautifulsoup4==4.8.2
billiard==3.6.3.0
celery==4.4.0
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
decorator==4.4.2
Django==2.2.10
django-extensions==2.2.8
django-rest-swagger==2.1.2
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
entrypoints==0.3
flake8==3.7.9
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.5.0
ipython==7.13.0
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.16.0
Jinja2==2.11.1
kombu==4.6.7
MarkupSafe==1.1.1
mccabe==0.6.1
openapi-codec==1.3.2
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
prompt-toolkit==3.0.3
psycopg2==2.8.4
ptyprocess==0.6.0
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.5.2
PyJWT==1.7.1
pytz==2019.3
requests==2.23.0
simplejson==3.17.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.0
traitlets==4.3.3
uritemplate==3.0.1
urllib3==1.25.8
vine==1.3.0
wcwidth==0.1.8
zipp==3.0.0
